requirements.txt
grpcio==1.62.1
grpcio-tools==1.62.1
pyyaml==6.0.1

tabulate